Functionally more than what we had with knockout (play pause visibility working). I feel that the view code with angular is cleaner. e.g: not one massive data-bind, no function calls () because of variable being observable.
The viewmodel however I am undecisive about. Angular has it all in $scope. Good news is that it forced me to use an interface in typescript which means that its VERY clear what is and what isn't being used from the view. (e.g. update is not defined on scope).